update the close and pdf functions
close changes the data at the time but does not persist the change to the local storage json object, but does confirm it works in the console logs. this also creates the pdf and displays the label and the information of everything on the actual form. if the form was open, it closes, if the form was closed, it stays closed